pyproject.toml
[project]
name = "ask-sg-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "fastapi>=0.115.6",
    "openai>=1.59.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
    "types-requests>=2.32.0.20241016",
    "uvicorn>=0.34.0",
]
